Add files via upload
Experimental firmware for Simon Templar. This firmware should shift the wireless copy frequency to 433.250 MHz. If you want to change this frequency you will need a hex editor.  
  
First decide on the frequency you want (must be UHF) in 10 Hertz units, so for example if you want to set it to 435.275Mhz this would be 43,527,500  
Now you need to convert this to Hex, so in this example that would be 02982D4C  
Now split this into 4 bytes, which would be for this example  02 98 2D 4C  
Now load the firmware into a hex editor and find the following offset bytes
F32D = 02  
F32F = 95  
F336 = 16  
F338 = 48  
Replace each of these offset bytes with the 4 bytes you previously calculated and this should set the Wireless Copy frequency to what you want.